api coupon list

+++ minimal-api-net8/Models/Coupon.cs
@@ -2,7 +2,7 @@
 {
     public class Coupon : BaseModel
     {
-        public decimal DiscountPercent { get; set; }
+        public double DiscountPercent { get; set; }
         public bool IsActive { get; set; }
         public DateTime EffectiveStartDate { get; set; }
         public DateTime EffectiveEndDate { get; set;}
